+++ R/export_mesh.R
@@ -160,8 +160,7 @@ plot_mesh <- function(
   suppressMessages(
     plot_cylinders(
       cylinder, radius, length, start_x, start_y, start_z,
-      axis_x, axis_y, axis_z, facets, colors,
-      axes = FALSE
+      axis_x, axis_y, axis_z, facets, colors, lit = TRUE
     )
   )